requirements.txt
mxnet==1.9.0
mxnet_cu102==1.6.0
numpy==1.18.1
nvidia_dali_cuda100==0.23.0
opencv_python==4.2.0.32
pandas==1.0.5
Pillow==9.0.0
scikit_learn==1.0.2
scipy==1.4.1
skimage==0.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.46.0
visdom==0.1.8.9
